requirements.txt
requests==2.24.0
jsonpath==0.82
Faker==1.0.7
PyMySQL==0.9.3
suds-jurko==0.6
pytest==4.5.0
allure-pytest==2.6.5
pytest-html==1.21.1
pytest-metadata==1.8.0
pytest-rerunfailures==7.0
pytest-parallel==0.1.1
selenium==3.141.0
pypiwin32==223
Click==7.0
